@@ -62,64 +44,70 @@
   "license": "MIT",
   "homepage": "https://github.com/thekevinscott/UpscalerJS",
   "peerDependencies": {
-    "@tensorflow/tfjs": "^3.13.0"
+    "@tensorflow/tfjs": "^3.18.0",
+    "@tensorflow/tfjs-node": "^3.18.0",
+    "@tensorflow/tfjs-node-gpu": "^3.18.0"
   },
   "devDependencies": {
-    "@babel/plugin-transform-modules-commonjs": "^7.16.0",
-    "@babel/preset-typescript": "^7.16.0",
-    "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-node-resolve": "^13.0.6",
-    "@tensorflow/tfjs": "^3.13.0",
-    "@tensorflow/tfjs-node": "^3.13.0",
-    "@tensorflow/tfjs-node-gpu": "^3.13.0",
-    "@types/inquirer": "^8.2.0",
-    "@types/jest": "^27.4.1",
+    "@babel/plugin-transform-modules-commonjs": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
+    "@rollup/plugin-commonjs": "^22.0.1",
+    "@rollup/plugin-node-resolve": "^13.3.0",
+    "@schemastore/package": "^0.0.6",
+    "@tensorflow/tfjs": "^3.18.0",
+    "@tensorflow/tfjs-node": "^3.18.0",
+    "@tensorflow/tfjs-node-gpu": "^3.18.0",
+    "@types/fs-extra": "^9.0.13",
+    "@types/jest": "^28.1.4",
     "@types/jsdom": "^16.2.14",
-    "@types/mocha": "^9.1.0",
+    "@types/mocha": "^9.1.1",
+    "@types/pixelmatch": "^5.2.4",
     "@types/pngjs": "^6.0.1",
-    "@types/puppeteer": "^5.4.4",
+    "@types/puppeteer": "^5.4.6",
     "@types/rimraf": "^3.0.2",
+    "@types/selenium-webdriver": "^4.1.1",
     "@types/serve-handler": "^6.1.1",
-    "@types/yargs": "^17.0.5",
-    "browserstack-local": "^1.4.8",
+    "browserstack-local": "^1.5.1",
     "chai": "^4.3.6",
     "chai-as-promised": "^7.1.1",
     "chokidar": "^3.5.3",
-    "codecov": "^3.7.2",
-    "docsify-cli": "^4.4.3",
-    "dotenv": "^8.2.0",
-    "esbuild": "^0.9.6",
+    "docsify-cli": "^4.4.4",
+    "dotenv": "^16.0.1",
+    "esbuild": "^0.11.0",
+    "fs-extra": "^10.1.0",
     "heapdump": "^0.3.15",
-    "html-webpack-plugin": "^5.3.0",
-    "inquirer": "^8.2.0",
-    "jest": "^27.5.1",
-    "jest-canvas-mock": "^2.2.0",
-    "jsdom": "^19.0.0",
-    "lerna": "^3.14.0",
-    "loader-utils": "^2.0.0",
-    "mocha": "^9.2.1",
-    "pixelmatch": "^5.2.1",
-    "playwright-test": "^7.2.2",
+    "html-webpack-plugin": "^5.5.0",
+    "jest": "^28.1.2",
+    "jest-canvas-mock": "^2.4.0",
+    "jsdom": "^20.0.0",
+    "mocha": "^10.0.0",
+    "pixelmatch": "^5.3.0",
+    "playwright-test": "^8.1.1",
     "pngjs": "^6.0.0",
-    "prettier": "^2.0.5",
-    "puppeteer": "^13.1.3",
+    "prettier": "^2.7.1",
+    "puppeteer": "^15.3.0",
     "rimraf": "^3.0.2",
-    "rollup": "^2.66.1",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "selenium-webdriver": "^4.0.0-beta.1",
-    "serve": "^11.3.2",
-    "tape": "^5.5.0",
+    "rollup": "^2.75.7",
+    "seedrandom": "~2.4.3",
+    "selenium-webdriver": "^4.3.0",
+    "serve": "^13.0.2",
+    "serve-handler": "^6.1.3",
+    "tape": "^5.5.3",
     "tree-kill": "^1.2.2",
-    "ts-jest": "^27.1.3",
-    "ts-node": "^10.6.0",
-    "tslint": "^6.1.2",
-    "tslint-config-prettier": "^1.18.0",
-    "tslint-to-eslint-config": "1.0.0",
-    "typescript": "^4.6.0",
-    "uglify": "^0.1.5",
-    "webpack": "^5.24.3",
-    "yargs": "^16.2.0"
+    "yargs": "^16.2.0",
+    "ts-jest": "^28.0.5",
+    "ts-node": "^10.8.2",
+    "typescript": "^4.7.4",
+    "uglify-js": "^3.16.2",
+    "upscaler": "workspace:*",
+    "webpack": "^5.73.0"
   },
-  "packageManager": "yarn@3.1.0"
+  "packageManager": "pnpm@3.1.0",
+  "pnpm": {
+    "overrides": {
+      "@tensorflow/tfjs": "3.18.0",
+      "@tensorflow/tfjs-node": "3.18.0",
+      "@tensorflow/tfjs-node-gpu": "3.18.0"
+    }
+  }
 }
