@@ -3,16 +3,13 @@ module.exports = {
     "browser": true,
     "es6": true
   },
+  "ignorePatterns": ["**/*.test.ts", "**/*.generated.ts"],
   "extends": [
     "plugin:@typescript-eslint/recommended",
     "plugin:@typescript-eslint/recommended-requiring-type-checking",
     "prettier",
   ],
   "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "project": "tsconfig.json",
-    "sourceType": "module"
-  },
   "plugins": [
     "@typescript-eslint"
   ],
@@ -26,5 +23,23 @@ module.exports = {
         "allowEmptyCatch": true
       }
     ],
-  }
+  },
+  overrides: [
+    {
+      files: ['*.ts'],
+
+      // As mentioned in the comments, you should extend TypeScript plugins here,
+      // instead of extending them outside the `overrides`.
+      // If you don't want to extend any rules, you don't need an `extends` attribute.
+      extends: [
+        'plugin:@typescript-eslint/recommended',
+        'plugin:@typescript-eslint/recommended-requiring-type-checking',
+      ],
+
+      "parserOptions": {
+        "project": "./tsconfig.eslint.json",
+        "sourceType": "module"
+      },
+    },
+  ],
 };
