@@ -1,3 +1,17 @@
-jest.setTimeout(5000);
-// silence tensorflow.js warnings
-console.warn = () => {};
+// // jest.setTimeout(5000);
+// // // silence tensorflow.js warnings
+// // console.warn = () => {};
+
+// import buildModels from "../../scripts/package-scripts/build-model";
+
+// // const setup = async () => {
+// //   await buildModels(['esrgan-slim',]);
+// // };
+
+// // export default setup;
+
+// const globalSetup = async function () {
+//   await buildModels(['esrgan-slim',]);
+// };
+
+// export default globalSetup;
