@@ -1,8 +1,7 @@
-import * as fs from 'fs';
-import * as tf from '@tensorflow/tfjs-node';
+import fs from 'fs';
+import { tf, } from './dependencies.generated';
 import { isFourDimensionalTensor, isThreeDimensionalTensor, isTensor, isString, } from './utils';
 
-/* eslint-disable @typescript-eslint/no-explicit-any */
 export const getInvalidTensorError = (input: tf.Tensor) => new Error(
   [
     `Unsupported dimensions for incoming pixels: ${input.shape.length}.`,
@@ -10,6 +9,7 @@ export const getInvalidTensorError = (input: tf.Tensor) => new Error(
   ].join(' '),
 );
 
+/* eslint-disable @typescript-eslint/no-explicit-any */
 export const getInvalidInput = (input: any) => new Error([
   `Unknown input ${JSON.stringify(input)} provided. Input must be either a rank 3 or 4 tensor,`,
   `a string representing a local path or http-accessible path to an image,`,
@@ -19,7 +19,7 @@ export const getInvalidInput = (input: any) => new Error([
 const isUint8Array = (input: GetImageAsTensorInput): input is Uint8Array => input.constructor === Uint8Array;
 const isBuffer = (input: GetImageAsTensorInput): input is Buffer => input.constructor === Buffer;
 
-const getTensorFromInput = async (input: GetImageAsTensorInput): Promise<tf.Tensor3D | tf.Tensor4D> => {
+const getTensorFromInput = (input: GetImageAsTensorInput): tf.Tensor3D | tf.Tensor4D => {
   if (isUint8Array(input)) {
     return tf.node.decodeImage(input);
   }
@@ -33,28 +33,29 @@ const getTensorFromInput = async (input: GetImageAsTensorInput): Promise<tf.Tens
   }
 
   if (isString(input)) {
-    if (input.startsWith('http')) {
-      const arrayBuffer = await fetch(input).then(r => r.blob()).then(blob => blob.arrayBuffer());
-      const image = new Uint8Array(arrayBuffer);
-      return tf.node.decodeImage(image);
-    } else {
+    // if (input.startsWith('http')) {
+    //   const arrayBuffer = await fetch(input).then(r => r.blob()).then(blob => blob.arrayBuffer());
+    //   const image = new Uint8Array(arrayBuffer);
+    //   return tf.node.decodeImage(image);
+    // } else {
       const image = new Uint8Array(fs.readFileSync(input));
       return tf.node.decodeImage(image);
-    }
+    // }
   }
 
-  throw getInvalidInput(input);
+  throw getInvalidInput(input); 
 };
 
 export type GetImageAsTensorInput = tf.Tensor3D | tf.Tensor4D | string | Uint8Array | Buffer;
+
+/* eslint-disable @typescript-eslint/require-await */
 export const getImageAsTensor = async (
   input: GetImageAsTensorInput,
 ): Promise<tf.Tensor4D> => {
-  const tensor = await getTensorFromInput(input);
+  const tensor = getTensorFromInput(input);
 
   if (isThreeDimensionalTensor(tensor)) {
-    /* eslint-disable @typescript-eslint/no-unnecessary-type-assertion */
-    const expandedTensor = tensor.expandDims(0) as tf.Tensor4D;
+    const expandedTensor: tf.Tensor4D = tensor.expandDims(0);
     tensor.dispose();
     return expandedTensor;
   }
