@@ -14,4 +14,5 @@
  * @module UpscalerJS
  */
 export { default, } from './upscaler';
-export { AbortError } from './upscale';
+export { AbortError, getRowsAndColumns, getTensorDimensions, } from './upscale';
+export type { ModelDefinition, SingleArgProgress, MultiArgProgress, Progress, } from './types';
