@@ -1,19 +1,27 @@
 import { Upscaler } from './upscaler';
-import * as loadModel from './loadModel';
+import * as loadModel from './loadModel.generated';
 import './warmup';
 import * as upscale from './upscale';
 jest.mock('./upscale', () => ({
-  ...jest.requireActual('./upscale'),
+  ...(jest.requireActual('./upscale') as typeof upscale),
 }));
-jest.mock('./loadModel');
+jest.mock('./loadModel.generated');
 jest.mock('./warmup');
 
 const mockedUpscale = upscale as jest.Mocked<typeof upscale>;
 const mockedLoadModel = loadModel as jest.Mocked<typeof loadModel>;
-(mockedLoadModel as any).default = async() => () => ({});
+mockedLoadModel.loadModel.mockImplementation(async () => {
+  return {
+    modelDefinition: {
+      path: 'foo',
+      scale: 2,
+    },
+    model: 'foo' as any,
+  };
+});
 
 describe('Upscaler', () => {
-  it('is able to abort multiple times', () => new Promise(async (resolve, reject) => {
+  it('is able to abort multiple times', (): Promise<void> => new Promise(async (resolve, reject) => {
     const tick = () => new Promise(resolve => setTimeout(resolve));
     let count = 0;
     const cancellableUpscale = jest.fn(async function (_1, _2, { signal }: {
