@@ -1,25 +1,28 @@
 /****
  * Tests that different build outputs all function correctly
  */
+import * as http from 'http';
 import { checkImage } from '../../lib/utils/checkImage';
 import { prepareScriptBundleForUMD, DIST as SCRIPT_DIST } from '../../lib/umd/prepare';
 import { startServer } from '../../lib/shared/server';
 import { prepareScriptBundleForESM, bundleWebpack, DIST as WEBPACK_DIST } from '../../lib/esm-webpack/prepare';
 import puppeteer from 'puppeteer';
+import * as tf from '@tensorflow/tfjs';
+import Upscaler, { ModelDefinition } from 'upscaler';
 
-const JEST_TIMEOUT_IN_SECONDS = 60;
+const JEST_TIMEOUT_IN_SECONDS = 120;
 jest.setTimeout(JEST_TIMEOUT_IN_SECONDS * 1000);
 jest.retryTimes(1);
 
 describe('Build Integration Tests', () => {
-  let server;
-  let browser;
-  let page;
+  let server: http.Server;
+  let browser: puppeteer.Browser;
+  let page: puppeteer.Page;
 
   const PORT = 8099;
 
   afterEach(async function afterEach() {
-    const stopServer = () => new Promise((resolve) => {
+    const stopServer = (): Promise<void | Error> => new Promise((resolve) => {
       if (server) {
         server.close(resolve);
       } else {
@@ -47,10 +50,27 @@ describe('Build Integration Tests', () => {
     const result = await page.evaluate(() => {
       const Upscaler = window['Upscaler'];
       const upscaler = new Upscaler({
-        model: '/pixelator/pixelator.json',
-        scale: 4,
+        model: {
+          path: '/pixelator/pixelator.json',
+          scale: 4,
+        },
       });
-      return upscaler.upscale(document.getElementById('flower'));
+      return upscaler.upscale(<HTMLImageElement>document.getElementById('flower'));
+    });
+    checkImage(result, "upscaled-4x-pixelator.png", 'diff.png');
+  });
+
+  it("upscales using a UMD build with a specified model", async () => {
+    await prepareScriptBundleForUMD();
+    server = await startServer(PORT, SCRIPT_DIST);
+    await startBrowser();
+    const result = await page.evaluate(() => {
+      const Upscaler = window['Upscaler'];
+      const pixelUpsampler = window['PixelUpsampler4x'];
+      const upscaler = new Upscaler({
+        model: pixelUpsampler,
+      });
+      return upscaler.upscale(<HTMLImageElement>document.getElementById('flower'));
     });
     checkImage(result, "upscaled-4x-pixelator.png", 'diff.png');
   });
@@ -64,11 +84,22 @@ describe('Build Integration Tests', () => {
     const result = await page.evaluate(() => {
       const Upscaler = window['Upscaler'];
       const upscaler = new Upscaler({
-        model: '/pixelator/pixelator.json',
-        scale: 4,
+        model: {
+          path: '/pixelator/pixelator.json',
+          scale: 4,
+        },
       });
       return upscaler.upscale(window['flower']);
     });
     checkImage(result, "upscaled-4x-pixelator.png", 'diff.png');
   });
 });
+
+declare global {
+  interface Window {
+    Upscaler: typeof Upscaler;
+    flower: string;
+    tf: typeof tf;
+    PixelUpsampler4x: ModelDefinition; 
+  }
+}
