@@ -1,43 +1,36 @@
-import * as fs from 'fs';
-import * as path from 'path';
-import * as esbuild from 'esbuild';
-import * as rimraf from 'rimraf';
+import fs from 'fs';
+import path from 'path';
+import { buildSync } from 'esbuild';
 import { copyFixtures } from '../utils/copyFixtures';
-import { updateTFJSVersion } from '../utils/updateTFJSVersion';
-import callExec from '../utils/callExec';
+import { installLocalPackages, installNodeModules } from '../shared/prepare';
+import { LOCAL_UPSCALER_NAME, LOCAL_UPSCALER_NAMESPACE } from './constants';
 
 const ROOT = path.join(__dirname);
 export const DIST = path.join(ROOT, '/dist');
-const NODE_MODULES = path.join(ROOT, '/node_modules');
 const UPSCALER_PATH = path.join(ROOT, '../../../packages/upscalerjs')
-
-const moveUpscalerToLocallyNamedPackage = async (localNameForPackage: string) => {
-  // Make sure we load the version local to node_modules, _not_ the local version on disk,
-  // so we can ensure the build process is accurate and working correctly
-  rimraf.sync(`${NODE_MODULES}/${localNameForPackage}`);
-
-  await callExec(`mkdir -p ./node_modules`, {
-    cwd: ROOT,
-  });
-
-  await callExec(`cp -r ${UPSCALER_PATH} ${NODE_MODULES}/${localNameForPackage}`, {
-    cwd: UPSCALER_PATH,
-  });
-  
-  const packageJSON = JSON.parse(fs.readFileSync(`${NODE_MODULES}/${localNameForPackage}/package.json`, 'utf-8'));
-  packageJSON.name = localNameForPackage;
-  fs.writeFileSync(`${NODE_MODULES}/${localNameForPackage}/package.json`, JSON.stringify(packageJSON, null, 2));
-}
+const MODELS_PATH = path.join(ROOT, '../../../models')
 
 export const bundle = async () => {
-  const localNameForPackage = 'upscaler-for-esbuild'
-  await updateTFJSVersion(ROOT);
-  await moveUpscalerToLocallyNamedPackage(localNameForPackage);
-  rimraf.sync(DIST);
+  await installNodeModules(ROOT);
+  await installLocalPackages(ROOT, [
+    {
+      src: UPSCALER_PATH,
+      name: LOCAL_UPSCALER_NAME,
+    },
+    {
+      src: path.resolve(MODELS_PATH, 'esrgan-legacy'),
+      name: path.join(LOCAL_UPSCALER_NAMESPACE, 'esrgan-legacy'),
+    },
+    {
+      src: path.resolve(MODELS_PATH, 'pixel-upsampler'),
+      name: path.join(LOCAL_UPSCALER_NAMESPACE, 'pixel-upsampler'),
+    },
+  ]);
   copyFixtures(DIST, false);
+
   const entryFiles = path.join(ROOT, 'src/index.js');
   try {
-    esbuild.buildSync({
+    buildSync({
       entryPoints: [entryFiles],
       bundle: true,
       loader: {
