@@ -1,5 +1,5 @@
 import * as tfjs from '@tensorflow/tfjs';
-import Upscaler from 'upscaler-for-webpack';
+import Upscaler from 'upscaler-for-esm-webpack';
 import flower from '../../../__fixtures__/flower-small.png';
 window.tfjs = tfjs;
 window.flower = flower;
