@@ -1,43 +1,42 @@
 import callExec from "../utils/callExec";
-import * as fs from 'fs';
-import * as path from 'path';
-import * as rimraf from 'rimraf';
+import { mkdirp } from "fs-extra";
+import fs from 'fs';
+import path from 'path';
+import crypto from 'crypto';
+import rimraf from 'rimraf';
+import { getCryptoName, installLocalPackages, installNodeModules } from "../shared/prepare";
+import { LOCAL_UPSCALER_NAME } from "./constants";
 
 const ROOT = path.join(__dirname);
-const NODE_MODULES = path.join(ROOT, '/node_modules');
 const UPSCALER_PATH = path.join(ROOT, '../../../packages/upscalerjs')
 
-const moveUpscalerToLocallyNamedPackage = async (localNameForPackage: string) => {
-  // Make sure we load the version local to node_modules, _not_ the local version on disk,
-  // so we can ensure the build process is accurate and working correctly
-  rimraf.sync(`${NODE_MODULES}/${localNameForPackage}`);
-
-  await callExec(`cp -r ${UPSCALER_PATH} ${NODE_MODULES}`, {
-    cwd: UPSCALER_PATH,
-  });
-
-  await callExec(`mv ${NODE_MODULES}/upscalerjs ${NODE_MODULES}/${localNameForPackage}`, {
-    cwd: UPSCALER_PATH,
-  });
-  
-  const packageJSON = JSON.parse(fs.readFileSync(`${NODE_MODULES}/${localNameForPackage}/package.json`, 'utf-8'));
-  packageJSON.name = localNameForPackage;
-  fs.writeFileSync(`${NODE_MODULES}/${localNameForPackage}/package.json`, JSON.stringify(packageJSON, null, 2));
-}
-
-export const prepareScriptBundleForCJS = async () => {
-  const localNameForPackage = 'upscaler-for-node';
-
-  await callExec('yarn', {
-    cwd: ROOT,
-  });
-
-  moveUpscalerToLocallyNamedPackage(localNameForPackage);
+export const prepareScriptBundleForNodeCJS = async () => {
+  await installNodeModules(ROOT);
+  await installLocalPackages(ROOT, [
+    {
+      src: UPSCALER_PATH,
+      name: LOCAL_UPSCALER_NAME,
+    },
+  ]);
 };
 
 type Stdout = (data: string) => void;
-export const executeNodeScript = async (file: string, stdout?: Stdout) => {
+export const executeNodeScriptFromFilePath = async (file: string, stdout?: Stdout) => {
   await callExec(`node "./src/${file}"`, {
     cwd: ROOT
   }, stdout);
+};
+
+export const executeNodeScript = async (contents: string, stdout?: Stdout) => {
+  const TMP = path.resolve(ROOT, './tmp');
+  await mkdirp(TMP);
+  const hash = getCryptoName(contents);
+  const FILENAME = path.resolve(TMP, `${hash}.js`);
+  fs.writeFileSync(FILENAME, contents, 'utf-8');
+
+  await callExec(`node "${FILENAME}"`, {
+    cwd: ROOT
+  }, stdout);
+
+  rimraf.sync(TMP);
 };
