@@ -1,6 +1,6 @@
 const { exec } = require("child_process");
 type StdOut = (chunk: string) => void;
-const callExec = (cmd: string, options: any, stdout?: StdOut | boolean) => new Promise((resolve, reject) => {
+const callExec = (cmd: string, options: any, stdout?: StdOut | boolean, stdErr: boolean = true): Promise<void> => new Promise((resolve, reject) => {
   const spawnedProcess = exec(cmd, options, (error: Error) => {
     if (error) {
       reject(error.message);
@@ -8,7 +8,10 @@ const callExec = (cmd: string, options: any, stdout?: StdOut | boolean) => new P
       resolve();
     }
   });
-  spawnedProcess.stderr.pipe(process.stderr);
+
+  if (stdErr) {
+    spawnedProcess.stderr.pipe(process.stderr);
+  }
 
   if (stdout === false) {
 
