@@ -1,5 +1,5 @@
-import * as fs from 'fs';
-import * as path from 'path';
+import fs from 'fs';
+import path from 'path';
 
 const ROOT = __dirname;
 const FIXTURES = path.join(ROOT, '../../__fixtures__');
