@@ -1,5 +1,5 @@
-import * as path from 'path';
-import * as fs from 'fs';
+import path from 'path';
+import fs from 'fs';
 import * as _PNG from 'pngjs';
 const { PNG } = _PNG;
 
