@@ -1,5 +1,5 @@
-import * as fs from 'fs';
-import * as path from 'path';
+import fs from 'fs';
+import path from 'path';
 import { getTFJSVersion } from '../utils/getTFJSVersion';
 import callExec from '../utils/callExec';
 
@@ -10,7 +10,7 @@ export const updateTFJSVersion = async (dirname: string) => {
   if (currentTFJSVersion !== packageJSON.dependencies['@tensorflow/tfjs']) {
     packageJSON.dependencies['@tensorflow/tfjs'] = getTFJSVersion();
     fs.writeFileSync(packageJSONPath, JSON.stringify(packageJSON, null, 2), 'utf-8');
-    await callExec('yarn', {
+    await callExec('pnpm', {
       cwd: dirname,
     });
   }
