@@ -1,7 +1,9 @@
-import puppeteer, { Browser, BrowserContext, Page, WaitTask } from 'puppeteer';
+import puppeteer, { Browser, BrowserContext, Page } from 'puppeteer';
 import { bundle, DIST } from '../../lib/esm-esbuild/prepare';
 import { startServer } from '../../lib/shared/server';
 import * as http from 'http';
+import Upscaler, { ModelDefinition } from 'upscaler';
+import * as tf from '@tensorflow/tfjs';
 
 const JEST_TIMEOUT_IN_SECONDS = 30;
 jest.setTimeout(JEST_TIMEOUT_IN_SECONDS * 1000);
@@ -10,7 +12,7 @@ jest.retryTimes(1);
 const EXPECTED_LAYER_MODELS = 2; // I don't know why, but we start with layer model references in memory.
 const EXPECTED_UPSCALERS = 0;
 
-const stopServer = (server?: http.Server) => new Promise((resolve) => {
+const stopServer = (server?: http.Server): Promise<void | undefined | Error> => new Promise((resolve) => {
   if (server) {
     server.close(resolve);
   } else {
@@ -20,7 +22,7 @@ const stopServer = (server?: http.Server) => new Promise((resolve) => {
 });
 
 // https://puppeteer.github.io/puppeteer/docs/10.0.0/puppeteer.page.queryobjects/#example
-const countObjects = async (page: Page, prototype: puppeteer.JSHandle) => {
+const countObjects = async (page: Page, prototype: puppeteer.JSHandle): Promise<number> => {
   const instances = await page.queryObjects(prototype);
   const numberOfObjects = await page.evaluate((i) => i.length, instances);
 
@@ -37,7 +39,7 @@ const PORT = 8099;
 const TIMES_TO_CHECK = 7;
 
 interface PrototypeDefinition {
-  name: string;
+  name: 'LayersModel' | 'Upscaler';
   prototype: (page: Page) => Promise<puppeteer.JSHandle>
 }
 interface TFJSMemory {
@@ -57,30 +59,24 @@ interface MemoryRecord {
   memory: TFJSMemory;
 }
 
-const getMemory = async (page: Page, prototypes: Array<PrototypeDefinition>): Promise<MemoryRecord> => {
+const getMemory = async (page: Page): Promise<MemoryRecord> => {
   const allObjects = await Promise.all(prototypes.map(async ({ prototype }) => {
     return countObjects(page, await prototype(page));
   }));
 
   const names = prototypes.map(p => p.name);
 
-  const tfMemory = await page.evaluate(() => window['tf'].memory());
+  // TODO: Type mismatch between TFJS exported types
+  const tfMemory: TFJSMemory = await page.evaluate(() => window['tf'].memory() as TFJSMemory);
 
-  return allObjects.reduce((obj, objects, i) => {
-    return {
-      ...obj,
-      [names[i]]: objects,
-    };
-  }, {
+  return allObjects.reduce((obj, objects, i) => ({
+    ...obj,
+    [names[i]]: objects,
+  }), {
     memory: tfMemory,
+    LayersModel: 0,
+    Upscaler: 0,
   });
-};
-
-const getStartingMemory = async (page: Page, prototypes: Array<PrototypeDefinition>) => {
-  const memory = await getMemory(page, prototypes);
-  expect(memory.LayersModel).toEqual(EXPECTED_LAYER_MODELS);
-  expect(memory.Upscaler).toEqual(EXPECTED_UPSCALERS);
-  return memory;
 };
 
 const prototypes: Array<PrototypeDefinition> = [
@@ -93,6 +89,13 @@ const prototypes: Array<PrototypeDefinition> = [
     name: 'Upscaler',
   },
 ];
+
+const getStartingMemory = async (page: Page, prototypes: Array<PrototypeDefinition>) => {
+  const memory = await getMemory(page);
+  expect(memory.LayersModel).toEqual(EXPECTED_LAYER_MODELS);
+  expect(memory.Upscaler).toEqual(EXPECTED_UPSCALERS);
+  return memory;
+};
 
 describe('Memory Leaks', () => {
   let browser: Browser;
@@ -128,20 +131,20 @@ describe('Memory Leaks', () => {
   })
 
   afterAll(async () => {
+    await new Promise(resolve => setTimeout(resolve, 10000))
     await Promise.all([
       stopServer(server),
       browser.close(),
     ]);
   });
 
-  const tick = async (dur = 10) => {
+  const tick = async (page: puppeteer.Page, tickTime = 10) => {
     await page.evaluate(async (duration) => {
-      const wait = () => new Promise(resolve => setTimeout(resolve, duration));
-      await wait();
-    }, dur);
+      await new Promise(resolve => setTimeout(resolve, duration));
+    }, tickTime);
   }
 
-  const checkMemory = (names: Array<string>, starting: MemoryRecord, ending: MemoryRecord) => {
+  const checkMemory = (names: Array<'LayersModel' | 'Upscaler'>, starting: MemoryRecord, ending: MemoryRecord) => {
     expect(starting.memory.numTensors).toEqual(ending.memory.numTensors);
     expect(starting.memory.numDataBuffers).toEqual(ending.memory.numDataBuffers);
     for (let i = 0; i < names.length; i++) {
@@ -152,7 +155,7 @@ describe('Memory Leaks', () => {
         expect(endingObjects).toEqual(startingObjects);
       } catch(err) {
         const diff = endingObjects - startingObjects;
-        expect(new Error(`Memory Leak, there are ${diff} objects of type ${name}.`)).toBeUndefined();
+        expect(new Error(`Memory Leak, there are ${diff} objects of type ${name} and there should be 0. Ending objects: ${endingObjects}, starting objects: ${startingObjects}`)).toBeUndefined();
       }
     }
 
@@ -193,7 +196,7 @@ describe('Memory Leaks', () => {
   // //   //       console.log(foo);
   // //   //     }, 1000)
   // //   //   }, TIMES_TO_CHECK);
-  // //   //   const endingMemory = await getMemory(page, prototypes);
+  // //   //   const endingMemory = await getMemory(page);
   // //   //   const names = prototypes.map(p => p.name);
   // //   //   checkMemory(names, startingMemory, endingMemory);
   // //   // });
@@ -211,7 +214,7 @@ describe('Memory Leaks', () => {
   // //   //       console.log(foo);
   // //   //     }, 1000)
   // //   //   }, TIMES_TO_CHECK);
-  // //   //   const endingMemory = await getMemory(page, prototypes);
+  // //   //   const endingMemory = await getMemory(page);
   // //   //   const names = prototypes.map(p => p.name);
   // //   //   checkMemory(names, startingMemory, endingMemory);
   // //   // });
@@ -225,7 +228,7 @@ describe('Memory Leaks', () => {
   //   //       t.dispose();
   //   //     }
   //   //   }, TIMES_TO_CHECK);
-  //   //   const endingMemory = await getMemory(page, prototypes);
+  //   //   const endingMemory = await getMemory(page);
   //   //   const names = prototypes.map(p => p.name);
   //   //   checkMemory(names, startingMemory, endingMemory);
   //   // });
@@ -237,12 +240,21 @@ describe('Memory Leaks', () => {
     await page.evaluate(async (times) => {
       const Upscaler = window['Upscaler'];
       for (let i = 0; i < times; i++) {
-        const upscaler = new Upscaler();
+        // TODO: Revert this to use the default
+        // const upscaler = new Upscaler();
+        const upscaler = new Upscaler({
+          model: {
+            path: '/pixelator/pixelator.json',
+            scale: 4,
+          }
+        });
         await upscaler.dispose();
       }
     }, TIMES_TO_CHECK);
 
-    const endingMemory = await getMemory(page, prototypes);
+    await tick(page);
+
+    const endingMemory = await getMemory(page);
     const names = prototypes.map(p => p.name);
     checkMemory(names, startingMemory, endingMemory);
   });
@@ -252,15 +264,23 @@ describe('Memory Leaks', () => {
 
     await page.evaluate(async (times) => {
       const Upscaler = window['Upscaler'];
+      const ESRGANGANS = window['esrgan-legacy']['gans'];
       for (let i = 0; i < times; i++) {
+        // TODO: Revert this to use the default
+        // const upscaler = new Upscaler({
+        //   warmupSizes: [[50, 50]],
+        // });
         const upscaler = new Upscaler({
-          warmupSizes: [50, 50],
+          warmupSizes: [[50, 50]],
+          model: ESRGANGANS,
         });
         await upscaler.dispose();
       }
     }, TIMES_TO_CHECK);
 
-    const endingMemory = await getMemory(page, prototypes);
+    await tick(page);
+
+    const endingMemory = await getMemory(page);
     const names = prototypes.map(p => p.name);
     checkMemory(names, startingMemory, endingMemory);
   });
@@ -272,19 +292,17 @@ describe('Memory Leaks', () => {
       const Upscaler = window['Upscaler'];
       for (let i = 0; i < times; i++) {
         const upscaler = new Upscaler({
-          model: '/pixelator/pixelator.json',
-          scale: 4,
+          model: {
+            path: '/pixelator/pixelator.json',
+            scale: 4,
+          },
         });
         await upscaler.dispose();
       }
     }, TIMES_TO_CHECK);
 
-    // give a tick to clean up
-    await page.evaluate(async (duration) => {
-      const wait = () => new Promise(resolve => setTimeout(resolve, duration));
-      await wait();
-    }, 10);
-    const endingMemory = await getMemory(page, prototypes);
+    await tick(page);
+    const endingMemory = await getMemory(page);
     const names = prototypes.map(p => p.name);
     checkMemory(names, startingMemory, endingMemory);
   });
@@ -298,8 +316,10 @@ describe('Memory Leaks', () => {
         let image;
         for (let i = 0; i < times; i++) {
           const upscaler = new Upscaler({
-            model: '/pixelator/pixelator.json',
-            scale: 4,
+            model: {
+              path: '/pixelator/pixelator.json',
+              scale: 4,
+            },
           });
           image = await upscaler.upscale(window['flower']);
 
@@ -308,8 +328,8 @@ describe('Memory Leaks', () => {
         return image;
       }, TIMES_TO_CHECK);
 
-      await tick();
-      const endingMemory = await getMemory(page, prototypes);
+      await tick(page);
+      const endingMemory = await getMemory(page);
       const names = prototypes.map(p => p.name);
       checkMemory(names, startingMemory, endingMemory);
       expect(image.substring(0,22)).toEqual('data:image/png;base64,');
@@ -324,9 +344,9 @@ describe('Memory Leaks', () => {
         let image;
         for (let i = 0; i < times; i++) {
           const upscaler = new Upscaler({
-            modelDefinition: {
+            model: {
               scale: 4,
-              url: '/pixelator/pixelator.json',
+              path: '/pixelator/pixelator.json',
               preprocess: (image) => tf.mul(image, 1),
             }
           });
@@ -337,8 +357,8 @@ describe('Memory Leaks', () => {
         return image;
       }, TIMES_TO_CHECK);
 
-      await tick();
-      const endingMemory = await getMemory(page, prototypes);
+      await tick(page);
+      const endingMemory = await getMemory(page);
       const names = prototypes.map(p => p.name);
       checkMemory(names, startingMemory, endingMemory);
       expect(image.substring(0,22)).toEqual('data:image/png;base64,');
@@ -353,9 +373,9 @@ describe('Memory Leaks', () => {
         let image;
         for (let i = 0; i < times; i++) {
           const upscaler = new Upscaler({
-            modelDefinition: {
+            model: {
               scale: 4,
-              url: '/pixelator/pixelator.json',
+              path: '/pixelator/pixelator.json',
               postprocess: (image) => tf.mul(image, 1),
             }
           });
@@ -366,8 +386,8 @@ describe('Memory Leaks', () => {
         return image;
       }, TIMES_TO_CHECK);
 
-      await tick();
-      const endingMemory = await getMemory(page, prototypes);
+      await tick(page);
+      const endingMemory = await getMemory(page);
       const names = prototypes.map(p => p.name);
       checkMemory(names, startingMemory, endingMemory);
       expect(image.substring(0,22)).toEqual('data:image/png;base64,');
@@ -382,9 +402,9 @@ describe('Memory Leaks', () => {
         let image;
         for (let i = 0; i < times; i++) {
           const upscaler = new Upscaler({
-            modelDefinition: {
+            model: {
               scale: 4,
-              url: '/pixelator/pixelator.json',
+              path: '/pixelator/pixelator.json',
               preprocess: (image) => tf.mul(image, 1),
               postprocess: (image) => tf.mul(image, 1),
             }
@@ -396,8 +416,8 @@ describe('Memory Leaks', () => {
         return image;
       }, TIMES_TO_CHECK);
 
-      await tick();
-      const endingMemory = await getMemory(page, prototypes);
+      await tick(page);
+      const endingMemory = await getMemory(page);
       const names = prototypes.map(p => p.name);
       checkMemory(names, startingMemory, endingMemory);
       expect(image.substring(0,22)).toEqual('data:image/png;base64,');
@@ -412,9 +432,9 @@ describe('Memory Leaks', () => {
       const Upscaler = window['Upscaler'];
       for (let i = 0; i < times; i++) {
         const upscaler = new Upscaler({
-          modelDefinition: {
+          model: {
             scale: 4,
-            url: '/pixelator/pixelator.json',
+            path: '/pixelator/pixelator.json',
             preprocess: (image) => tf.mul(image, 1),
             postprocess: (image) => tf.mul(image, 1),
           }
@@ -429,15 +449,15 @@ describe('Memory Leaks', () => {
       }
     }, TIMES_TO_CHECK);
 
-    await tick();
-    const endingMemory = await getMemory(page, prototypes);
+    await tick(page);
+    const endingMemory = await getMemory(page);
     const names = prototypes.map(p => p.name);
     checkMemory(names, startingMemory, endingMemory);
   });
 
   it('should upscale with a pre and a post processing functions from a tensor', async () => {
     await page.evaluate(async () => {
-      const getImage = () => new Promise(resolve => {
+      const getImage = (): Promise<HTMLImageElement> => new Promise(resolve => {
         const img = new Image();
         img.src = window['flower'];
         img.crossOrigin = 'anonymous';
@@ -451,12 +471,12 @@ describe('Memory Leaks', () => {
     const image = await page.evaluate(async (times) => {
       const tf = window['tf'];
       const Upscaler = window['Upscaler'];
-      let output;
+      let output: string;
       for (let i = 0; i < times; i++) {
         const upscaler = new Upscaler({
-          modelDefinition: {
+          model: {
             scale: 4,
-            url: '/pixelator/pixelator.json',
+            path: '/pixelator/pixelator.json',
             preprocess: (image) => tf.mul(image, 1),
             postprocess: (image) => tf.mul(image, 1),
           }
@@ -468,8 +488,8 @@ describe('Memory Leaks', () => {
       return output;
     }, TIMES_TO_CHECK);
 
-    await tick();
-    const endingMemory = await getMemory(page, prototypes);
+    await tick(page);
+    const endingMemory = await getMemory(page);
     const names = prototypes.map(p => p.name);
     checkMemory(names, startingMemory, endingMemory);
     expect(image.substring(0,22)).toEqual('data:image/png;base64,');
@@ -485,9 +505,9 @@ describe('Memory Leaks', () => {
       let output;
       for (let i = 0; i < times; i++) {
         const upscaler = new Upscaler({
-          modelDefinition: {
+          model: {
             scale: 4,
-            url: '/pixelator/pixelator.json',
+            path: '/pixelator/pixelator.json',
             preprocess: (image) => tf.mul(image, 1),
             postprocess: (image) => tf.mul(image, 1),
           }
@@ -501,8 +521,8 @@ describe('Memory Leaks', () => {
       return output;
     }, TIMES_TO_CHECK);
 
-    await tick();
-    const endingMemory = await getMemory(page, prototypes);
+    await tick(page);
+    const endingMemory = await getMemory(page);
     const names = prototypes.map(p => p.name);
     checkMemory(names, startingMemory, endingMemory);
     expect(image.substring(0,22)).toEqual('data:image/png;base64,');
@@ -511,12 +531,12 @@ describe('Memory Leaks', () => {
   it('should upscale with the idealo model', async () => {
     const startingMemory = await getStartingMemory(page, prototypes);
     const image = await page.evaluate(async (times) => {
-      const tf = window['tf'];
       const Upscaler = window['Upscaler'];
+      const ESRGANGANS = window['esrgan-legacy']['gans'];
       let output;
       for (let i = 0; i < times; i++) {
         const upscaler = new Upscaler({
-          model: 'idealo/gans',
+          model: ESRGANGANS,
         });
         output = await upscaler.upscale(window['flower']);
 
@@ -525,8 +545,8 @@ describe('Memory Leaks', () => {
       return output;
     }, TIMES_TO_CHECK);
 
-    await tick();
-    const endingMemory = await getMemory(page, prototypes);
+    await tick(page);
+    const endingMemory = await getMemory(page);
     const names = prototypes.map(p => p.name);
     checkMemory(names, startingMemory, endingMemory);
     expect(image.substring(0,22)).toEqual('data:image/png;base64,');
@@ -539,8 +559,10 @@ describe('Memory Leaks', () => {
       let output;
       for (let i = 0; i < times; i++) {
         const upscaler = new Upscaler({
-          model: '/pixelator/pixelator.json',
-          scale: 4,
+          model: {
+            path: '/pixelator/pixelator.json',
+            scale: 4,
+          },
         });
         await upscaler.upscale(window['flower'], {
           output: 'src',
@@ -556,22 +578,24 @@ describe('Memory Leaks', () => {
       return output;
     }, TIMES_TO_CHECK);
 
-    await tick();
-    const endingMemory = await getMemory(page, prototypes);
+    await tick(page);
+    const endingMemory = await getMemory(page);
     const names = prototypes.map(p => p.name);
     checkMemory(names, startingMemory, endingMemory);
-    expect(image.substring(0,22)).toEqual('data:image/png;base64,');
+    expect((image as string).substring(0,22)).toEqual('data:image/png;base64,');
   });
 
   it('should callback to progress with a tensor', async () => {
     const startingMemory = await getStartingMemory(page, prototypes);
     const image = await page.evaluate(async (times) => {
       const Upscaler = window['Upscaler'];
-      let output;
+      let output: tf.Tensor;
       for (let i = 0; i < times; i++) {
         const upscaler = new Upscaler({
-          model: '/pixelator/pixelator.json',
-          scale: 4,
+          model: {
+            path: '/pixelator/pixelator.json',
+            scale: 4,
+          },
         });
         await upscaler.upscale(window['flower'], {
           output: 'src',
@@ -592,12 +616,12 @@ describe('Memory Leaks', () => {
       return output;
     }, TIMES_TO_CHECK);
 
-    await tick();
+    await tick(page);
     expect(image.shape).toEqual([8, 8, 3]);
     await page.evaluate(() => {
       window['output'].dispose();
     })
-    const endingMemory = await getMemory(page, prototypes);
+    const endingMemory = await getMemory(page);
     const names = prototypes.map(p => p.name);
     checkMemory(names, startingMemory, endingMemory);
   });
@@ -609,8 +633,10 @@ describe('Memory Leaks', () => {
       const abortController = new AbortController();
       for (let i = 0; i < times; i++) {
         const upscaler = new Upscaler({
-          model: '/pixelator/pixelator.json',
-          scale: 4,
+          model: {
+            path: '/pixelator/pixelator.json',
+            scale: 4,
+          },
         });
         upscaler.upscale(window['flower'], {
           output: 'src',
@@ -622,8 +648,8 @@ describe('Memory Leaks', () => {
       }
     }), TIMES_TO_CHECK);
 
-    await tick();
-    const endingMemory = await getMemory(page, prototypes);
+    await tick(page);
+    const endingMemory = await getMemory(page);
     const names = prototypes.map(p => p.name);
     checkMemory(names, startingMemory, endingMemory);
   });
@@ -635,8 +661,10 @@ describe('Memory Leaks', () => {
       const abortController = new AbortController();
       for (let i = 0; i < times; i++) {
         const upscaler = new Upscaler({
-          model: '/pixelator/pixelator.json',
-          scale: 4,
+          model: {
+            path: '/pixelator/pixelator.json',
+            scale: 4,
+          },
         });
         try {
           await upscaler.upscale(window['flower'], {
@@ -644,7 +672,7 @@ describe('Memory Leaks', () => {
             signal: abortController.signal,
             patchSize: 14,
             padding: 2,
-            progress: (rate) => {
+            progress: (rate: number) => {
               if (rate >= .5) {
                 abortController.abort();
               }
@@ -656,8 +684,8 @@ describe('Memory Leaks', () => {
       }
     }, TIMES_TO_CHECK);
 
-    await tick();
-    const endingMemory = await getMemory(page, prototypes);
+    await tick(page);
+    const endingMemory = await getMemory(page);
     const names = prototypes.map(p => p.name);
     checkMemory(names, startingMemory, endingMemory);
   });
@@ -669,8 +697,10 @@ describe('Memory Leaks', () => {
       const abortController = new AbortController();
       for (let i = 0; i < times; i++) {
         const upscaler = new Upscaler({
-          model: '/pixelator/pixelator.json',
-          scale: 4,
+          model: {
+            path: '/pixelator/pixelator.json',
+            scale: 4,
+          },
         });
         try {
           await upscaler.upscale(window['flower'], {
@@ -679,7 +709,7 @@ describe('Memory Leaks', () => {
             patchSize: 14,
             padding: 2,
             progress: (rate, slice) => {
-              slice.dispose();
+              (slice as unknown as tf.Tensor).dispose();
               if (rate >= .5) {
                 abortController.abort();
               }
@@ -691,9 +721,21 @@ describe('Memory Leaks', () => {
       }
     }, TIMES_TO_CHECK);
 
-    await tick();
-    const endingMemory = await getMemory(page, prototypes);
+    await tick(page);
+    const endingMemory = await getMemory(page);
     const names = prototypes.map(p => p.name);
     checkMemory(names, startingMemory, endingMemory);
   });
 });
+
+declare global {
+  interface Window {
+    Upscaler: typeof Upscaler;
+    flower: string;
+    tf: typeof tf;
+    pixelUpsampler: ModelDefinition;
+    src?: tf.Tensor4D | tf.Tensor3D;
+    output?: tf.Tensor;
+    'esrgan-legacy': Record<string, ModelDefinition>;
+  }
+}
